// frontend/lib/api/generated/model/imageMetadata.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * kartonko API
 * kartonko web server
 * OpenAPI spec version: 1.0
 */

export interface ImageMetadata {
  filename: string;
  format: string;
  hash: string;
  height: number;
  id: number;
  tags: string[];
  uploaded_at: string;
  user_id: number;
  width: number;
}